refactor: survey/create-card -> survey/create/card
