+++ src/components/surveys/create/card/CreateSection.jsx
@@ -1,6 +1,6 @@
 import React from "react";
-import { Create, Summary, Title } from "../common/styled";
-import { surveyTitle, surveySummary, surveyListState } from "../../../atoms";
+import { Create, Summary, Title } from "../../common/styled";
+import { surveyTitle, surveySummary, surveyListState } from "../../../../atoms";
 import { useRecoilValue, useSetRecoilState } from "recoil";
 import SurveyCreateItem from "./SurveyCreateItem";
 import SurveyItemCreator from "./SurveyItemCreator";
